% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{feor93_08_crosswalk}
\alias{feor93_08_crosswalk}
\title{FEOR-93 to FEOR-08 crosswalk}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 817 rows and 2 columns.
}
\usage{
feor93_08_crosswalk
}
\description{
FEOR-93 to FEOR-08 crosswalk
}
\keyword{datasets}
