% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{feor93_2015_05_19}
\alias{feor93_2015_05_19}
\title{FEOR-93}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 653 rows and 12 columns.
}
\usage{
feor93_2015_05_19
}
\description{
FEOR-93
}
\keyword{datasets}
