% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{isco08_2017_07_24}
\alias{isco08_2017_07_24}
\title{ISCO08}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 436 rows and 8 columns.}
\usage{
isco08_2017_07_24
}
\description{
ISCO08
}
\keyword{datasets}
