% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nace_r2_2018_06_06}
\alias{nace_r2_2018_06_06}
\title{NACE}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 615 rows and 10 columns.
}
\usage{
nace_r2_2018_06_06
}
\description{
NACE
}
\keyword{datasets}
