% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{teaor08_2018_09_01}
\alias{teaor08_2018_09_01}
\title{TEAOR08}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 615 rows and 10 columns.
}
\usage{
teaor08_2018_09_01
}
\description{
TEAOR08
}
\keyword{datasets}
